Moved Value into a sub namespace.
